Acceptable results

+++ src/Fomore/Core/Training/Neuro/WeightMatrix.cs
@@ -6,7 +6,7 @@ namespace Core.Training.Neuro
     [Serializable]
     public struct WeightMatrix
     {
-        private float[,] Weights { get; }
+        private double[,] Weights { get; }
 
         /// <summary>
         ///     Number of columns depends on the from layer.
@@ -18,18 +18,18 @@ namespace Core.Training.Neuro
         /// </summary>
         public int RowCount { get; }
 
-        public float this[int column, int row] => Weights[column, row];
+        public double this[int column, int row] => Weights[column, row];
 
-        public WeightMatrix(float[,] weights)
+        public WeightMatrix(double[,] weights)
         {
             Weights = weights;
             ColumnCount = weights.GetLength(0);
             RowCount = weights.GetLength(1);
         }
 
-        public float[,] GetClonedWeights()
+        public double[,] GetClonedWeights()
         {
-            var weights = new float [Weights.GetLength(0), Weights.GetLength(1)];
+            var weights = new double [Weights.GetLength(0), Weights.GetLength(1)];
             for (int i = 0; i < weights.GetLength(0); i++)
             {
                 for (int j = 0; j < weights.GetLength(1); j++)
@@ -46,7 +46,7 @@ namespace Core.Training.Neuro
             if (!(obj is WeightMatrix)) return false;
 
             var matrix = (WeightMatrix)obj;
-            return EqualityComparer<float[,]>.Default.Equals(Weights, matrix.Weights) &&
+            return EqualityComparer<double[,]>.Default.Equals(Weights, matrix.Weights) &&
                    ColumnCount == matrix.ColumnCount &&
                    RowCount == matrix.RowCount;
         }
@@ -54,7 +54,7 @@ namespace Core.Training.Neuro
         public override int GetHashCode()
         {
             int hashCode = -1291767658;
-            hashCode = hashCode * -1521134295 + EqualityComparer<float[,]>.Default.GetHashCode(Weights);
+            hashCode = hashCode * -1521134295 + EqualityComparer<double[,]>.Default.GetHashCode(Weights);
             hashCode = hashCode * -1521134295 + ColumnCount.GetHashCode();
             hashCode = hashCode * -1521134295 + RowCount.GetHashCode();
             return hashCode;
@@ -64,9 +64,9 @@ namespace Core.Training.Neuro
 
         public static bool operator !=(WeightMatrix matrix1, WeightMatrix matrix2) => !(matrix1 == matrix2);
 
-        public static WeightMatrix operator *(float scalar, WeightMatrix matrix)
+        public static WeightMatrix operator *(double scalar, WeightMatrix matrix)
         {
-            var weights = new float[matrix.ColumnCount, matrix.RowCount];
+            var weights = new double[matrix.ColumnCount, matrix.RowCount];
             for (int i = 0; i < matrix.ColumnCount; i++)
             {
                 for (int j = 0; j < matrix.RowCount; j++) weights[i, j] = matrix[i, j] * scalar;
@@ -75,9 +75,9 @@ namespace Core.Training.Neuro
             return new WeightMatrix(weights);
         }
 
-        public static WeightMatrix operator *(WeightMatrix matrix, float scalar)
+        public static WeightMatrix operator *(WeightMatrix matrix, double scalar)
         {
-            var weights = new float[matrix.ColumnCount, matrix.RowCount];
+            var weights = new double[matrix.ColumnCount, matrix.RowCount];
             for (int i = 0; i < matrix.ColumnCount; i++)
             {
                 for (int j = 0; j < matrix.RowCount; j++) weights[i, j] = matrix[i, j] * scalar;
@@ -86,9 +86,9 @@ namespace Core.Training.Neuro
             return new WeightMatrix(weights);
         }
 
-        public static float[] operator *(float[] values, WeightMatrix m)
+        public static double[] operator *(double[] values, WeightMatrix m)
         {
-            var newWeights = new float[m.RowCount];
+            var newWeights = new double[m.RowCount];
 
             for (int i = 0; i < m.RowCount; i++)
             {
